Test_code/requirements.txt
langchain
langchain_openai
langchain-community
python-dotenv
pypdf
chromadb
streamlit
